timestamp data type changed from unsigned int to long long, seaport despawn

+++ packet.h
@@ -56,7 +56,7 @@ typedef struct _LWPTTLSTATICSTATE2 {
     unsigned char padding0;
     unsigned char padding1;
     unsigned char padding2;
-    unsigned int ts;
+    long long ts;
     int xc0;
     int yc0;
     int view_scale;
@@ -78,7 +78,7 @@ typedef struct _LWPTTLSEAPORTSTATE {
     unsigned char padding0;
     unsigned char padding1;
     unsigned char padding2;
-    unsigned int ts;
+    long long ts;
     int xc0;
     int yc0;
     int view_scale;
@@ -156,7 +156,7 @@ typedef struct _LWPTTLPINGCHUNK {
     unsigned char padding1;
     unsigned char padding2;
     int chunk_key;
-    unsigned int ts;
+    long long ts;
 } LWPTTLPINGCHUNK;
 /*
 * END: should sync with packet.h in sea-server
